Adding additional endpoints
Added a few additional API endpoints: ability to run portfolio compliance, and ability to test a rule. The UI needs these functional to be able to have all endpoints to be built. Yes, I'm being lazy and just keeping these changes in feature/streamlit when ideally they should be in a separate branch. :)